+++ kroxylicious-filters/kroxylicious-sasl-inspection/src/test/java/io/kroxylicious/filter/sasl/inspection/SaslInspectionFilterTest.java
@@ -336,7 +336,6 @@ class SaslInspectionFilterTest {
                 .satisfies(rfr -> assertThat(rfr.message())
                         .isEqualTo(expectedDownstreamAuthenticateShortCircuitResponse));
 
-        verify(context, never()).clientSaslAuthenticationSuccess(anyString(), anyString());
     }
 
     @Test
@@ -379,7 +378,6 @@ class SaslInspectionFilterTest {
                     assertThat(rfr.closeConnection()).isTrue();
                 });
 
-        verify(context, never()).clientSaslAuthenticationSuccess(anyString(), anyString());
     }
 
     @Test
@@ -484,7 +482,6 @@ class SaslInspectionFilterTest {
                     assertThat(rfr.closeConnection()).isTrue();
                 });
 
-        verify(context, never()).clientSaslAuthenticationSuccess(anyString(), anyString());
         verify(context).clientSaslAuthenticationFailure(eq("PLAIN"), eq("tim"), isA(SaslException.class));
     }
 
@@ -635,7 +632,6 @@ class SaslInspectionFilterTest {
         doSaslAuthenticateRequest(TestData.SASL_OAUTHBEARER_SIGNED_JWT, filter, (short) 0);
         doSaslAuthenticateResponse(new byte[0], filter, 0 /* a session lifespan of 0ms means the client must re-auth next */);
 
-        verify(context, never()).clientSaslAuthenticationSuccess(anyString(), anyString());
         verify(context).clientSaslAuthenticationFailure(eq("OAUTHBEARER"), eq("johndoe"), isA(SaslException.class));
     }
 
@@ -680,7 +676,7 @@ class SaslInspectionFilterTest {
 
         // Then
         verify(context).clientSaslAuthenticationFailure(eq(observerFactory.mechanismName()), eq(expectedAuthorizedId), isA(SubjectBuildingException.class));
-        verify(context, never()).clientSaslAuthenticationSuccess(anyString(), anyString());
+
         verify(context, never()).clientSaslAuthenticationSuccess(anyString(), any(Subject.class));
         // verify(context).r
 
@@ -712,7 +708,7 @@ class SaslInspectionFilterTest {
 
         // Then
         verify(context, never()).clientSaslAuthenticationSuccess(any(), any(Subject.class));
-        verify(context, never()).clientSaslAuthenticationSuccess(any(), any(String.class));
+
         verify(context, never()).clientSaslAuthenticationFailure(anyString(), anyString(), nullable(Exception.class));
         verify(context, never()).forwardRequest(any(), ArgumentMatchers.assertArg(r -> assertThat(ApiKeys.forId(r.apiKey())).isEqualTo(ApiKeys.METADATA)));
         verify(requestCloseOrTerminalStage).withCloseConnection();
@@ -733,7 +729,7 @@ class SaslInspectionFilterTest {
 
         // Then
         verify(context, never()).clientSaslAuthenticationSuccess(any(), any(Subject.class));
-        verify(context, never()).clientSaslAuthenticationSuccess(any(), any(String.class));
+
         verify(context, never()).clientSaslAuthenticationFailure(anyString(), anyString(), nullable(Exception.class));
         verify(context, times(1)).forwardRequest(any(), ArgumentMatchers.assertArg(r -> assertThat(ApiKeys.forId(r.apiKey())).isEqualTo(ApiKeys.METADATA)));
         assertThat(filterResult)
@@ -766,7 +762,7 @@ class SaslInspectionFilterTest {
 
         // Then
         verify(context, never()).clientSaslAuthenticationSuccess(any(), any(Subject.class));
-        verify(context, never()).clientSaslAuthenticationSuccess(any(), any(String.class));
+
         verify(context, never()).clientSaslAuthenticationFailure(anyString(), anyString(), nullable(Exception.class));
         var inOrder = Mockito.inOrder(context);
         inOrder.verify(context).forwardRequest(any(), ArgumentMatchers.argThat(r -> ApiKeys.forId(r.apiKey()).equals(ApiKeys.SASL_HANDSHAKE)));